+++ internal/domain/user.go
@@ -10,7 +10,7 @@ type User struct {
 type UserService interface {
 	Register(ctx context.Context, user *User, uniqueLoginErrorChan chan error) error
 	CompareHashAndPassword(ctx context.Context, user *User) (bool, error)
-	AddOrder(ctx context.Context, orderNumber int) error
+	AddOrder(ctx context.Context, orderNumber int64) error
 	ReadOrders(ctx context.Context) ([]Order, error)
 	ReadBalance(ctx context.Context) (Balance, error)
 }
@@ -19,7 +19,7 @@ type UserService interface {
 type UserRepository interface {
 	Register(ctx context.Context, user User, uniqueLoginErrorChan chan error) error
 	GetPasswordHash(ctx context.Context, login string) (string, error)
-	AddOrder(ctx context.Context, orderNumber int) error
+	AddOrder(ctx context.Context, orderNumber int64) error
 	ReadOrders(ctx context.Context) ([]Order, error)
 	ReadBalance(ctx context.Context) (Balance, error)
 }
