+++ pkg/migrations/001_init.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 BEGIN TRANSACTION;
-CREATE TABLE IF NOT EXISTS orders(num INTEGER PRIMARY KEY, uploaded_at TIMESTAMP, stat TEXT, username TEXT);
+CREATE TABLE IF NOT EXISTS orders(num BIGINT PRIMARY KEY, uploaded_at TIMESTAMP, stat TEXT, username TEXT);
 CREATE TABLE IF NOT EXISTS balances(username TEXT PRIMARY KEY, balance INTEGER, withdrawn INTEGER);
 COMMIT;
